feat: adopt @use syntax to support new Sass module system
