+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Change Log
 ## [Unreleased]
 
+## [3.0.0] - 2020-01-10
+* Supports the new Sass module system released with Dart Sass 1.23.0. The use of @use is recommended instead of @import.
+
 ## [2.0.0] - 2018-11-06
 * Puts back in reverted changes from 1.3.4, support for Webpack v4 and no support for Webpack v1. [#66](https://github.com/shakacode/sass-resources-loader/pull/66) by [david-ogalego](https://github.com/david-ogalego)
 
@@ -24,6 +27,7 @@
 For other releases see [GitHub Releases section](https://github.com/shakacode/sass-resources-loader/releases).
 
 [Unreleased]: https://github.com/shakacode/sass-resources-loader/compare/v2.0.0...master
+[2.0.0]: https://github.com/shakacode/sass-resources-loader/compare/v2.0.0...v3.0.0
 [2.0.0]: https://github.com/shakacode/sass-resources-loader/compare/v1.3.5...v2.0.0
 [1.3.5]: https://github.com/shakacode/sass-resources-loader/compare/v1.3.4...v1.3.5
 [1.3.4]: https://github.com/shakacode/sass-resources-loader/compare/v1.3.3...v1.3.4
