+++ test/scss/shared/_index.scss
@@ -1,0 +1,1 @@
+@forward "shared/variables";
